+++ __init__.py
@@ -1,4 +1,4 @@
-from .addon import register
+from .goal_tracking_progress_bar.addon import register
 
 
 register()